Clarifying the tiled/untiled creation options with comment, new test
Adding comments to the tiled creation option test.  Tiled typically means that the block
size is square, but it's only the GDAL GeoTiff implementation that defaults to 256x256.

Adding a new test for when we set TILED=NO for the creation options.  This forces a raster
to be accessed in a strip fashion, row-by-row.